Usunused code cleanup
Remove some warnings, move tests to new location.
